refactor ( client ): refactor folder structure
