+++ stack/client/src/layouts/splash/SectionAbout.tsx
