+++ stack/client/src/layouts/splash/SectionCapabilities.tsx
